DS-630: Organized tests in to 'development' and 'released' folders. 'released' is currently supported by Ed Analytics - 'development' is not supported yet.
