+++ src/test/java/org/imsglobal/caliper/ri/events/RIConstants.java
@@ -1,7 +1,9 @@
 package org.imsglobal.caliper.ri.events;
 
 public class RIConstants {
-	
+
     public static final String APP_NAME = "ReadingInventory";
+    public static final String RIMI_RELEASED_DIRECTORY = "fixtures/hmh/hmh-ri/released/";
+    public static final String RIMI_DEVELOPMENT_DIRECTORY = "fixtures/hmh/hmh-ri/development/";
 
 }
